+++ flecsi/topo/ntree/interface.hh
@@ -84,32 +84,17 @@ public:
   // tree.
   ntree(const coloring & c)
     : with_meta<Policy>(c.nparts_),
-      part{{
-        make_repartitioned<Policy, entities>(c.nparts_,
-          make_partial<allocate>(c.entities_sizes_)),
-        make_repartitioned<Policy, nodes>(c.nparts_,
-          make_partial<allocate>(c.nodes_sizes_)),
-        make_repartitioned<Policy, hashmap>(c.nparts_,
-          make_partial<allocate>(
-            std::vector<util::id>(c.nparts_, c.local_hmap_))),
-        make_repartitioned<Policy, tree_data>(c.nparts_,
-          make_partial<allocate>(std::vector<util::id>(c.nparts_, 3))),
-        make_repartitioned<Policy, comms>(c.nparts_,
-          make_partial<allocate>(std::vector<util::id>(c.nparts_, c.nparts_))),
-        make_repartitioned<Policy, share_ghosts_comms>(c.nparts_,
-          make_partial<allocate>(
-            std::vector<util::id>(c.nparts_, c.nparts_ * 2 - 1))),
-        make_repartitioned<Policy, top_tree_ents>(c.nparts_,
-          make_partial<allocate>(std::vector<util::id>(c.nparts_, 1))),
-        make_repartitioned<Policy, top_tree_nodes>(c.nparts_,
-          make_partial<allocate>(std::vector<util::id>(c.nparts_, 1))),
-        make_repartitioned<Policy, share_ghosts_cid_comm>(c.nparts_,
-          make_partial<allocate>(std::vector<util::id>(c.nparts_, 1))),
-        make_repartitioned<Policy, share_ghosts_buffer_comm>(c.nparts_,
-          make_partial<allocate>(std::vector<util::id>(c.nparts_, 1))),
-        make_repartitioned<Policy, share_ghosts_distant_buffer_comm>(c.nparts_,
-          make_partial<allocate>(std::vector<util::id>(c.nparts_, 1))),
-      }},
+      part{{rep<entities>(c, c.entities_sizes_),
+        rep<nodes>(c, c.nodes_sizes_),
+        rep<hashmap>(c, c.local_hmap_),
+        rep<tree_data>(c, 3),
+        rep<comms>(c, c.nparts_),
+        rep<share_ghosts_comms>(c, c.nparts_ * 2 - 1),
+        rep<top_tree_ents>(c, 1),
+        rep<top_tree_nodes>(c, 1),
+        rep<share_ghosts_cid_comm>(c, 1),
+        rep<share_ghosts_buffer_comm>(c, 1),
+        rep<share_ghosts_distant_buffer_comm>(c, 1)}},
       cp_data_tree(*this,
         // Avoid initializer-list constructor:
         data::copy_plan::Sizes(c.nparts_, 1),
@@ -118,11 +103,11 @@ public:
         util::constant<tree_data>()),
       cp_share_ghosts_comms(*this,
         // Avoid initializer-list constructor:
-        data::copy_plan::Sizes(c.nparts_, c.nparts_),
+        data::copy_plan::Sizes(c.nparts_, 1),
         task<set_dests_share_ghosts_comms>,
         task<set_ptrs_share_ghosts_comms>,
         util::constant<share_ghosts_comms>()),
-      buf([c] {
+      buf([&c] {
         data::buffers::coloring ret(c.nparts_);
         for(std::size_t i_r = 0; i_r < ret.size(); ++i_r) {
           for(std::size_t i = 0; i < c.nparts_; ++i) {
@@ -135,6 +120,19 @@ public:
       }()) {
     // Initialize the meta_field
     flecsi::execute<init_meta_field>(meta_field(this->meta), c.entities_sizes_);
+  }
+
+private:
+  template<index_space idx>
+  auto rep(const coloring & c, const std::vector<util::id> & size) {
+    return make_repartitioned<Policy, idx>(
+      c.nparts_, make_partial<allocate>(size));
+  }
+
+  template<index_space idx>
+  auto rep(const coloring & c, util::id size) {
+    return make_repartitioned<Policy, idx>(
+      c.nparts_, make_partial<allocate_same>(size));
   }
 
   // Ntree mandatory fields ---------------------------------------------------
@@ -211,8 +209,6 @@ private:
 
   // Buffer for ghosts shared
   data::buffers::core buf;
-  static constexpr util::id buffer_size =
-    (data::buffers::Buffer::size / sizeof(entity_data)) * 2;
 
   /// Hashing table type
   using hmap_t = util::hashtable<ntree::key_t, ntree::hcell_t, Policy>;
@@ -223,10 +219,9 @@ private:
   }
 
   static void init_meta_field(
-    typename field<meta_type, data::single>::template accessor<wo> meta_field,
-    const std::vector<util::id> size) {
-    meta_field = {};
-    meta_field->local.ents = size[run::context::instance().color()];
+    typename field<meta_type, data::single>::template accessor<wo> mf,
+    const std::vector<util::id> & size) {
+    mf->local.ents = size[run::context::instance().color()];
   }
 
   // ----------------------- Top Tree Construction Tasks -----------------------
@@ -239,7 +234,7 @@ private:
     typename field<key_t>::template accessor<rw, na> n_keys,
     typename field<ntree_data>::template accessor<ro, na> data_field,
     typename field<hmap_pair_t>::template accessor<rw, na> hcells,
-    typename field<meta_type, data::single>::template accessor<rw> meta_field) {
+    typename field<meta_type, data::single>::template accessor<rw> mf) {
     // Cstr htable
     auto hmap = map(hcells);
 
@@ -257,7 +252,7 @@ private:
     const auto lobound = color == 0 ? key_t::min() : data_field(1).hibound;
     // Check sort and data_field communication
     flog_assert(lobound <= e_keys(0), "The keys are not globally sorted");
-    flog_assert(hibound >= e_keys(meta_field->local.ents - 1),
+    flog_assert(hibound >= e_keys(mf->local.ents - 1),
       "The keys are not globally sorted");
 
     // Add the root
@@ -265,7 +260,7 @@ private:
     auto root_ = hmap.find(key_t::root());
     root_->second.set_color(color);
     {
-      const std::size_t cnode = meta_field->local.nodes++;
+      const std::size_t cnode = mf->local.nodes++;
       root_->second.set_node_idx(cnode);
       n_keys(cnode) = root_->second.key();
     }
@@ -356,8 +351,6 @@ private:
       // Prepare next loop
       lastekey = ekey;
       lastnkey = nkey;
-      meta_field->max_depth = std::max(meta_field->max_depth, current_depth);
-
     } // for
 
     // Generate the indices of the local nodes
@@ -370,7 +363,7 @@ private:
       auto nkey = cur->key();
       if(cur->key() != key_t::root()) {
         assert(cur->idx() == 0);
-        std::size_t cnode = meta_field->local.nodes++;
+        std::size_t cnode = mf->local.nodes++;
         cur->set_node_idx(cnode);
         n_keys(cnode) = cur->key();
       }
@@ -401,11 +394,7 @@ private:
     T & count_accessor_ents,
     T & count_accessor_nodes) {
 
-    std::optional<std::size_t> count_ents, count_nodes;
-    if constexpr(!C) {
-      count_ents = 0;
-      count_nodes = 0;
-    }
+    [[maybe_unused]] std::size_t count_ents = 0, count_nodes = 0;
 
     auto color = run::context::instance().color();
     std::vector<hcell_t *> queue;
@@ -428,13 +417,13 @@ private:
             if constexpr(C)
               ++count_accessor_nodes;
             else
-              count_accessor_nodes[(*count_nodes)++] = *cur;
+              count_accessor_nodes[count_nodes++] = *cur;
           }
           else {
             if constexpr(C)
               ++count_accessor_ents;
             else
-              count_accessor_ents[(*count_ents)++] = *cur;
+              count_accessor_ents[count_ents++] = *cur;
           }
         } // else
       } // for
@@ -450,12 +439,12 @@ private:
     typename key_t::int_t child,
     const int & color,
     hmap_t & hmap,
-    typename field<meta_type, data::single>::template accessor<rw> meta_field,
+    typename field<meta_type, data::single>::template accessor<rw> mf,
     typename field<key_t>::template accessor<rw, na> n_keys) {
     auto parent = hmap.end();
     while((parent = hmap.find(key)) == hmap.end()) {
       parent = hmap.insert(key, key);
-      const std::size_t cnode = meta_field->local.nodes++;
+      const std::size_t cnode = mf->local.nodes++;
       parent->second.set_node_idx(cnode);
       n_keys(cnode) = key;
       parent->second.add_child(child);
@@ -470,7 +459,7 @@ private:
   static void load_shared_entity(const std::size_t & c,
     const key_t & k,
     hmap_t & hmap,
-    typename field<meta_type, data::single>::template accessor<rw> meta_field,
+    typename field<meta_type, data::single>::template accessor<rw> mf,
     typename field<key_t>::template accessor<rw, na> n_keys) {
     auto key = k;
     auto f = hmap.find(key);
@@ -478,11 +467,11 @@ private:
       auto & cur = hmap.insert(key, key)->second;
       cur.set_nonlocal();
       cur.set_color(c);
-      auto eid = meta_field->local.ents + meta_field->ghosts++;
+      auto eid = mf->local.ents + mf->ghosts++;
       cur.set_ent_idx(eid);
       // Add missing parent(s)
       auto lastbit = key.pop();
-      add_parent(key, lastbit, c, hmap, meta_field, n_keys);
+      add_parent(key, lastbit, c, hmap, mf, n_keys);
     }
     else {
       assert(false);
@@ -492,7 +481,7 @@ private:
   static void load_shared_node(const std::size_t & c,
     const key_t & k,
     hmap_t & hmap,
-    typename field<meta_type, data::single>::template accessor<rw> meta_field,
+    typename field<meta_type, data::single>::template accessor<rw> mf,
     typename field<key_t>::template accessor<rw, na> n_keys) {
     key_t key = k;
     // Node doesnt exists already
@@ -503,7 +492,7 @@ private:
       cur.set_color(c);
       // Add missing parent(s)
       auto lastbit = key.pop();
-      add_parent(key, lastbit, c, hmap, meta_field, n_keys);
+      add_parent(key, lastbit, c, hmap, mf, n_keys);
     }
     else {
       assert(false);
@@ -515,7 +504,7 @@ private:
   // The new sizes are then returned to create the copy plan for the top tree.
   static void make_tree_distributed_task(
     typename field<key_t>::template accessor<rw, na> n_keys,
-    typename field<meta_type, data::single>::template accessor<rw> meta_field,
+    typename field<meta_type, data::single>::template accessor<rw> mf,
     typename field<ntree_data>::template accessor<rw, na> data_field,
     typename field<hmap_pair_t>::template accessor<rw, na> hcell,
     data::multi<typename field<hcell_t>::template accessor<ro, na>>
@@ -531,7 +520,7 @@ private:
         if(c.color() == color)
           continue;
         ++total_ents;
-        load_shared_entity(c.color(), c.key(), hmap, meta_field, n_keys);
+        load_shared_entity(c.color(), c.key(), hmap, mf, n_keys);
       }
     }
     for(auto & m : nodes_cells_multi.accessors()) {
@@ -539,7 +528,7 @@ private:
         if(c.color() == color)
           continue;
         ++total_nodes;
-        load_shared_node(c.color(), c.key(), hmap, meta_field, n_keys);
+        load_shared_node(c.color(), c.key(), hmap, mf, n_keys);
       }
     }
     // Update the lo and hi bounds
@@ -551,20 +540,18 @@ private:
       for(auto & c : m.span()) {
         if(c.color() == color)
           continue;
-        assert(n_keys.span().size() >
-               meta_field->local.nodes + meta_field->top_tree + 1);
+        assert(n_keys.span().size() > mf->local.nodes + mf->top_tree + 1);
         auto cur = hmap.find(c.key());
         assert(cur != hmap.end());
-        const std::size_t cnode =
-          meta_field->local.nodes + meta_field->top_tree++;
+        const std::size_t cnode = mf->local.nodes + mf->top_tree++;
         cur->second.set_node_idx(cnode);
         n_keys(cnode) = cur->second.key();
       }
     }
 
     // Update meta for copy plans
-    meta_field->cp_nents_tt = total_ents;
-    meta_field->cp_nnodes_tt = total_nodes;
+    mf->cp_nents_tt = total_ents;
+    mf->cp_nnodes_tt = total_nodes;
   }
 
   static void increase_size_task(topo::resize::Field::accessor<rw> a,
@@ -574,20 +561,18 @@ private:
 
   static void resize_entities_update_meta_task(
     topo::resize::Field::accessor<rw> a,
-    typename field<meta_type, data::single>::template accessor<rw> m) {
-    m->nents_base = m->local.ents;
-    m->nents_tt = m->buffer_counter;
-    a += m->nents_base + m->nents_tt;
+    typename field<meta_type, data::single>::template accessor<rw> mf) {
+    a = mf->local.ents + mf->nents_recv_2;
   }
 
   static void copy_sizes_meta_top_tree_task(topo::resize::Field::accessor<wo> a,
-    typename field<meta_type, data::single>::template accessor<ro> m) {
-    a = m->local.ents + m->cp_nents_tt;
+    typename field<meta_type, data::single>::template accessor<ro> mf) {
+    a = mf->local.ents + mf->cp_nents_tt;
   }
 
   static void copy_sizes_meta_task(topo::resize::Field::accessor<wo> a,
-    typename field<meta_type, data::single>::template accessor<ro> m) {
-    a = m->local.ents;
+    typename field<meta_type, data::single>::template accessor<ro> mf) {
+    a = mf->local.ents;
   }
 
   static void copy_sizes_task(topo::resize::Field::accessor<wo> a,
@@ -599,9 +584,10 @@ private:
 
   template<bool C>
   static void copy_sizes_resize_task(topo::resize::Field::accessor<rw> a,
+    typename field<meta_type, data::single>::template accessor<ro> mf,
     field<util::id>::accessor<ro, na> b) {
     const auto color = run::context::instance().color();
-    a = (C ? a : 0) + b[color] +
+    a = (C ? mf->local.ents : 0) + b[color] +
         std::accumulate(b.span().begin() + run::context::instance().colors(),
           b.span().end(),
           0);
@@ -610,52 +596,59 @@ private:
   template<index_space E = entities>
   static void copy_sizes_top_tree_task(topo::resize::Field::accessor<wo> a,
     future<std::array<std::size_t, 2>> b) {
-    a = b.get()[E == entities ? 0 : 1];
+    a = b.get()[E != entities];
   }
 
   // Return the number of entities/nodes for local, top tree and ghosts
   // stores in the respective index spaces.
   static std::array<util::id, 4> sizes_task(
-    typename field<meta_type, data::single>::template accessor<ro> m) {
-    return {{m->local.ents, m->local.nodes, m->top_tree, m->ghosts}};
+    typename field<meta_type, data::single>::template accessor<ro> mf) {
+    return {{mf->local.ents, mf->local.nodes, mf->top_tree, mf->ghosts}};
   }
 
   static void reset_ghosts(
-    typename field<meta_type, data::single>::template accessor<rw> m) {
-    m->ghosts = 0;
+    typename field<meta_type, data::single>::template accessor<rw> mf,
+    typename field<hmap_pair_t>::template accessor<rw, na> hcell) {
+    auto hmap = map(hcell);
+    // Remove ghosts from hmap
+    for(auto & i : hmap)
+      if(i.second.is_ent() && i.second.is_nonlocal()) {
+        auto c = i.first.pop();
+        hmap.find(i.first)->second.remove_child(c);
+        i = {};
+      }
+    mf->ghosts = 0;
   }
 
   // Copy plan: set destination sizes
   template<index_space IS = entities>
   static void set_destination_meta_top_tree(
     field<data::intervals::Value>::accessor<wo> a,
-    typename field<meta_type, data::single>::template accessor<ro> m) {
-    auto i = run::context::instance().color();
-    if constexpr(IS == entities)
-      a(0) = data::intervals::make(
-        {m->local.ents, m->local.ents + m->cp_nents_tt}, i);
-    else
-      a(0) = data::intervals::make(
-        {m->local.nodes, m->local.nodes + m->cp_nnodes_tt}, i);
+    typename field<meta_type, data::single>::template accessor<ro> mf) {
+    const auto n = IS == entities ? mf->local.ents : mf->local.nodes;
+    a(0) = data::intervals::make(
+      {n, n + (IS == entities ? mf->cp_nents_tt : mf->cp_nnodes_tt)},
+      run::context::instance().color());
   }
 
   // Copy plan: set destination sizes
   static void set_destination_meta(
     field<data::intervals::Value>::accessor<wo> a,
-    typename field<meta_type, data::single>::template accessor<ro> m) {
-    a(0) = data::intervals::make({m->nents_base, m->nents_base + m->nents_tt},
-      run::context::instance().color());
+    typename field<meta_type, data::single>::template accessor<ro> mf) {
+    a(0) =
+      data::intervals::make({mf->local.ents, mf->local.ents + mf->nents_recv_2},
+        run::context::instance().color());
   }
 
   // Copy plan: set pointers for top tree
   template<index_space IS = entities>
   static void set_top_tree_ptrs(
     field<data::copy_engine::Point>::accessor<wo, wo> a,
-    typename field<meta_type, data::single>::template accessor<ro> m,
+    typename field<meta_type, data::single>::template accessor<ro> mf,
     data::multi<typename field<hcell_t>::template accessor<ro, na>>
       hcells_multi) {
-    auto i = run::context::instance().color();
-    util::id idx = IS == entities ? m->local.ents : m->local.nodes;
+    const auto i = run::context::instance().color();
+    util::id idx = IS == entities ? mf->local.ents : mf->local.nodes;
     for(auto & m : hcells_multi.accessors())
       for(auto & c : m.span())
         if((IS == entities ? c.is_ent() : c.is_node()) && c.color() != i)
@@ -665,10 +658,10 @@ private:
   // Copy plan: set pointers to entities
   static void set_entities_ptrs(
     field<data::copy_engine::Point>::accessor<wo, wo> a,
-    typename field<meta_type, data::single>::template accessor<ro> m,
+    typename field<meta_type, data::single>::template accessor<ro> mf,
     typename field<h_s_t>::template accessor<ro, na> ids) {
-    util::id idx = m->nents_base;
-    for(std::size_t j = 0; j < m->buffer_counter; ++j) {
+    util::id idx = mf->local.ents;
+    for(std::size_t j = 0; j < mf->nents_recv_2; ++j) {
       assert(ids[j].first.color() != run::context::instance().color());
       a(idx++) =
         data::copy_engine::point(ids[j].first.color(), ids[j].first.idx());
@@ -679,22 +672,20 @@ private:
   // This shares part of the top tree information.
   static void exchange_boundaries_task(
     typename field<key_t>::template accessor<ro, na> e_keys,
-    typename field<meta_type, data::single>::template accessor<rw> m,
+    typename field<meta_type, data::single>::template accessor<rw> mf,
     typename field<ntree_data>::template accessor<rw, na> df) {
     auto c = run::context::instance().color();
-    m->max_depth = 0;
-    m->local.ents = e_keys.span().size();
+    mf->local.ents = e_keys.span().size();
     df(0).lobound = c == 0 ? key_t::min() : e_keys(0);
     df(0).hibound = c == run::context::instance().colors() - 1
                       ? key_t::max()
-                      : e_keys(m->local.ents - 1);
+                      : e_keys(mf->local.ents - 1);
   }
 
   // Recolor the entities, some might have moved after the sort.
   static void recolor_task(typename Policy::template accessor<rw, wo> t) {
-    for(auto & v : t.e_colors.span()) {
+    for(auto & v : t.e_colors.span())
       v = run::context::instance().color();
-    }
   }
 
   // ---------------------------- Top tree construction -----------------------
@@ -714,13 +705,17 @@ public:
     auto fm_top_tree = flecsi::execute<make_tree_local_task>(
       e_keys(ts), n_keys(ts), data_field(ts), hcells(ts), meta_field(ts->meta));
 
-    execute<copy_sizes_top_tree_task<entities>>(
-      ts->template get_partition<top_tree_ents>().sizes(), fm_top_tree);
-    ts->template get_partition<top_tree_ents>().resize();
+    {
+      auto & p = ts->template get_partition<top_tree_ents>();
+      execute<copy_sizes_top_tree_task<entities>>(p.sizes(), fm_top_tree);
+      p.resize();
+    }
 
-    execute<copy_sizes_top_tree_task<nodes>>(
-      ts->template get_partition<top_tree_nodes>().sizes(), fm_top_tree);
-    ts->template get_partition<top_tree_nodes>().resize();
+    {
+      auto & p = ts->template get_partition<top_tree_nodes>();
+      execute<copy_sizes_top_tree_task<nodes>>(p.sizes(), fm_top_tree);
+      p.resize();
+    }
 
     flecsi::execute<fill_top_tree_task>(
       hcells(ts), top_tree_ents_field(ts), top_tree_nodes_field(ts));
@@ -736,9 +731,12 @@ public:
       top_tree_ents_field(lm_top_tree),
       top_tree_nodes_field(lm_top_tree));
 
-    flecsi::execute<copy_sizes_meta_top_tree_task>(
-      ts->template get_partition<entities>().sizes(), meta_field(ts->meta));
-    ts->template get_partition<entities>().resize();
+    {
+      auto & p = ts->template get_partition<entities>();
+      flecsi::execute<copy_sizes_meta_top_tree_task>(
+        p.sizes(), meta_field(ts->meta));
+      p.resize();
+    }
 
     // Fake initialization for the new ghosts
     for(auto & f :
@@ -783,15 +781,14 @@ private:
     typename field<color_id>::accessor<ro, na> f) {
     std::fill(restart.span().begin(), restart.span().end(), 0);
     util::id cur = 0;
+    const auto color = run::context::instance().color();
     for(Color c = 0; c < run::context::instance().colors(); ++c) {
-      if(c != run::context::instance().color()) {
+      if(c != color) {
         auto w = mv[cur].write();
         for(std::size_t i = 0; i < f.span().size(); ++i) {
-          if(f[i].color == c) {
-            if(!w(a(f[i].id))) {
-              restart(cur) = i;
-              break;
-            } // if
+          if(f[i].color == c && !w(a(f[i].id))) {
+            restart(cur) = i;
+            break;
           } // if
         } // for
         ++cur;
@@ -809,11 +806,9 @@ private:
       if(c != color) {
         auto w = mv[cur].write();
         for(std::size_t i = 0; i < f.span().size(); ++i) {
-          if(f[i].second == c) {
-            if(!w(f[i])) {
-              restart(cur) = i;
-              break;
-            } // if
+          if(f[i].second == c && !w(f[i])) {
+            restart(cur) = i;
+            break;
           } // if
         } // for
         ++cur;
@@ -845,7 +840,7 @@ private:
       } // if
     } // for
 
-    // Keep copy if needed
+    // Keep copying if needed
     bool more_to_copy = false;
     cur = 0;
     for(Color c = 0; c < cs; ++c) {
@@ -854,13 +849,11 @@ private:
         if(restart(cur) != 0) {
           auto w = mv[cur].write();
           for(std::size_t i = restart(c); i < f.span().size(); ++i) {
-            if(f[i].color == c) {
-              if(!w(a(f[i].id))) {
-                restart(cur) = i;
-                done = false;
-                more_to_copy = true;
-                break;
-              }
+            if(f[i].color == c && !w(a(f[i].id))) {
+              restart(cur) = i;
+              done = false;
+              more_to_copy = true;
+              break;
             } // if
           } // for
           if(done)
@@ -886,13 +879,13 @@ private:
       if(c != color) {
         auto r = mv[cur + cs - 1].read();
         while(r) {
-          a(m->buffer_counter++) = r();
+          a(m->nents_recv_2++) = r();
         } // while
         ++cur;
       } // if
     } // for
 
-    // Keep copy if needed
+    // Keep copying if needed
     bool more_to_copy = false;
     cur = 0;
     for(Color c = 0; c < cs; ++c) {
@@ -901,13 +894,11 @@ private:
         if(restart(cur) != 0) {
           auto w = mv[cur].write();
           for(std::size_t i = restart(c); i < f.span().size(); ++i) {
-            if(f[i].second == c) {
-              if(!w(f[i])) {
-                restart(cur) = i;
-                done = false;
-                more_to_copy = true;
-                break;
-              }
+            if(f[i].second == c && !w(f[i])) {
+              restart(cur) = i;
+              done = false;
+              more_to_copy = true;
+              break;
             } // if
           } // for
           if(done)
@@ -954,18 +945,18 @@ private:
 
   static void load_entities_task(
     typename field<hmap_pair_t>::template accessor<rw, na> hcells,
-    typename field<meta_type, data::single>::template accessor<rw> meta_field,
+    typename field<meta_type, data::single>::template accessor<rw> mf,
     typename field<h_s_t>::template accessor<ro, na> recv) {
     auto hmap = map(hcells);
     auto c = run::context::instance().color();
-    for(std::size_t i = 0; i < meta_field->buffer_counter; ++i) {
+    for(std::size_t i = 0; i < mf->nents_recv_2; ++i) {
       auto key = recv[i].first.key();
       auto f = hmap.find(key);
       if(f == hmap.end()) {
         auto & cur = hmap.insert(key, key)->second;
         cur.set_nonlocal();
         cur.set_color(c);
-        auto eid = meta_field->local.ents + meta_field->ghosts++;
+        auto eid = mf->local.ents + mf->ghosts++;
         cur.set_ent_idx(eid);
         // Add missing parent(s)
         auto lastbit = key.pop();
@@ -973,7 +964,7 @@ private:
       }
       else {
         auto & cur = hmap.insert(key, key)->second;
-        auto eid = meta_field->local.ents + meta_field->ghosts++;
+        auto eid = mf->local.ents + mf->ghosts++;
         cur.set_nonlocal();
         cur.set_color(c);
         cur.set_ent_idx(eid);
@@ -989,25 +980,28 @@ public:
 
     // Remove copy plan
     ts->cp_entities.reset();
-    // Reset ghosts and get current sizes
-    flecsi::execute<reset_ghosts>(meta_field(ts->meta));
 
     // Find entities that will be used
     flecsi::execute<find_local_task<true>>(ts, share_ghosts_comms_field(ts));
 
-    execute<copy_sizes_task>(
-      ts->template get_partition<share_ghosts_cid_comm>().sizes(),
-      share_ghosts_comms_field(ts));
-    ts->template get_partition<share_ghosts_cid_comm>().resize();
+    {
+      auto & p = ts->template get_partition<share_ghosts_cid_comm>();
+      execute<copy_sizes_task>(p.sizes(), share_ghosts_comms_field(ts));
+      p.resize();
+    }
 
     flecsi::execute<find_local_task<false>>(
       ts, share_ghosts_cid_comm_field(ts));
-    ts->cp_share_ghosts_comms.issue_copy(share_ghosts_comms_field.fid);
+    ts->cp_share_ghosts_comms.issue_copy({share_ghosts_comms_field.fid});
 
-    flecsi::execute<copy_sizes_resize_task<true>>(
-      ts->template get_partition<entities>().sizes(),
-      share_ghosts_comms_field(ts));
-    ts->template get_partition<entities>().resize();
+    flecsi::execute<reset_ghosts>(meta_field(ts->meta), hcells(ts));
+
+    {
+      auto & p = ts->template get_partition<entities>();
+      flecsi::execute<copy_sizes_resize_task<true>>(
+        p.sizes(), meta_field(ts->meta), share_ghosts_comms_field(ts));
+      p.resize();
+    }
 
     execute<xfer_entities_req_start>(
       e_i(ts), comms_field(ts), *(ts->buf), share_ghosts_cid_comm_field(ts));
@@ -1024,21 +1018,23 @@ public:
     flecsi::execute<find_distant_task<true>>(ts, share_ghosts_comms_field(ts));
 
     // Resize share_ghosts_buffer_comm_field
-    execute<copy_sizes_task>(
-      ts->template get_partition<share_ghosts_buffer_comm>().sizes(),
-      share_ghosts_comms_field(ts));
-    ts->template get_partition<share_ghosts_buffer_comm>().resize();
-
+    {
+      auto & p = ts->template get_partition<share_ghosts_buffer_comm>();
+      execute<copy_sizes_task>(p.sizes(), share_ghosts_comms_field(ts));
+      p.resize();
+    }
     // Now fill info
     flecsi::execute<find_distant_task<false>>(
       ts, share_ghosts_buffer_comm_field(ts));
-    ts->cp_share_ghosts_comms.issue_copy(share_ghosts_comms_field.fid);
+    ts->cp_share_ghosts_comms.issue_copy({share_ghosts_comms_field.fid});
 
     // Resize
-    execute<copy_sizes_resize_task<false>>(
-      ts->template get_partition<share_ghosts_distant_buffer_comm>().sizes(),
-      share_ghosts_comms_field(ts));
-    ts->template get_partition<share_ghosts_distant_buffer_comm>().resize();
+    {
+      auto & p = ts->template get_partition<share_ghosts_distant_buffer_comm>();
+      execute<copy_sizes_resize_task<false>>(
+        p.sizes(), meta_field(ts->meta), share_ghosts_comms_field(ts));
+      p.resize();
+    }
 
     // Perform buffered copy
     execute<xfer_entities_cp_start>(
@@ -1049,17 +1045,20 @@ public:
       comms_field(ts),
       *(ts->buf),
       share_ghosts_buffer_comm_field(ts))
-            .get()) {
-    } // while
+            .get())
+      ;
 
     // Load entities destinated for this rank
     flecsi::execute<load_entities_task>(hcells(ts),
       meta_field(ts->meta),
       share_ghosts_distant_buffer_comm_field(ts));
 
-    flecsi::execute<resize_entities_update_meta_task>(
-      ts->template get_partition<entities>().sizes(), meta_field(ts->meta));
-    ts->template get_partition<entities>().resize();
+    {
+      auto & p = ts->template get_partition<entities>();
+      flecsi::execute<resize_entities_update_meta_task>(
+        p.sizes(), meta_field(ts->meta));
+      p.resize();
+    }
 
     // create copy plan for ghosts entities
     auto entities_dests_task = [&ts](auto f) {
@@ -1097,7 +1096,6 @@ private:
     typename field<node_data>::template accessor<wo, na> ni) {
     hmap_t hmap(hm.span());
     hmap.clear();
-    mf->max_depth = 0;
     mf->local.nodes = 0;
     mf->top_tree = 0;
     mf->ghosts = 0;
@@ -1114,11 +1112,10 @@ public:
     ts->cp_entities.reset();
 
     // Resize the entities
+    auto & p = ts.get().template get_partition<entities>();
     std::vector<util::id> nents_rz(ts->colors());
-    flecsi::execute<copy_sizes_meta_task>(
-      ts.get().template get_partition<entities>().sizes(),
-      meta_field(ts->meta));
-    ts.get().template get_partition<entities>().resize();
+    flecsi::execute<copy_sizes_meta_task>(p.sizes(), meta_field(ts->meta));
+    p.resize();
   }
 
   //---------------------------------------------------------------------------
@@ -1182,7 +1179,7 @@ struct ntree<Policy>::access {
 private:
   accessor<ntree::data_field> data_field;
   accessor<ntree::hcells> hcells;
-  data::scalar_access<ntree::meta_field, privilege_pack<ro>> meta_field;
+  data::scalar_access<ntree::meta_field, privilege_pack<ro>> mf;
 
 public:
   template<class F>
@@ -1195,7 +1192,7 @@ public:
     hcells.topology_send(f);
     e_i.topology_send(f);
     n_i.topology_send(f);
-    meta_field.topology_send(
+    mf.topology_send(
       std::forward<F>(f), [](auto & n) -> auto & { return n.meta; });
   }
 
@@ -1230,16 +1227,14 @@ public:
       hcell_t * cur = tqueue.front();
       tqueue.pop();
       // Intersection
-      if(f(cur)) {
-        if(cur->has_child()) {
-          auto nkey = cur->key();
-          for(std::size_t j = 0; j < nchildren_; ++j) {
-            if(cur->has_child(j)) {
-              tqueue.push(&hmap.at(nkey.push(j)));
-            } // if
-          } // for
-        }
-      }
+      if(f(cur) && cur->has_child()) {
+        auto nkey = cur->key();
+        for(std::size_t j = 0; j < nchildren_; ++j) {
+          if(cur->has_child(j)) {
+            tqueue.push(&hmap.at(nkey.push(j)));
+          } // if
+        } // for
+      } // if
     } // while
   }
 
@@ -1247,16 +1242,12 @@ public:
   void find_intersect_entities(T count_accessor) const {
     auto hmap = map();
     const auto cs = run::context::instance().colors();
-    std::optional<std::size_t> count;
-    if constexpr(C)
-      assert(count_accessor.span().size() == cs * 2 - 1);
-    else
-      count = 0;
+    [[maybe_unused]] std::size_t count = 0;
     // Make a tree traversal per last elements in the intersection field.
     // Caution entities can be detected several time for the same neighbor.
     std::vector<std::set<hcell_t>> send_ids(cs);
-    std::size_t start = meta_field->local.ents;
-    std::size_t stop = start + meta_field->nents_recv;
+    std::size_t start = mf->local.ents;
+    std::size_t stop = start + mf->nents_recv;
     for(std::size_t i = start; i < stop; ++i) {
       ent_id id(i);
       auto tcolor = e_colors[i];
@@ -1285,8 +1276,8 @@ public:
         count_accessor[i] = send_ids[i].size();
       else {
         for(auto a : send_ids[i]) {
-          assert(*count < count_accessor.span().size());
-          count_accessor[(*count)++] = {a, i};
+          assert(count < count_accessor.span().size());
+          count_accessor[count++] = {a, i};
         }
       }
     }
@@ -1301,7 +1292,7 @@ public:
       count = 0;
 
     auto hmap = map();
-    for(std::size_t i = 0; i < meta_field->local.ents; ++i) {
+    for(std::size_t i = 0; i < mf->local.ents; ++i) {
       std::set<std::size_t> send_colors;
       ent_id id(i);
       traversal(
@@ -1348,12 +1339,12 @@ public:
   FLECSI_INLINE_TARGET auto entities() const {
     if constexpr(PT == ptype_t::exclusive) {
       return make_ids<index_space::entities>(
-        util::iota_view<util::id>(0, meta_field->local.ents));
+        util::iota_view<util::id>(0, mf->local.ents));
     }
     else if constexpr(PT == ptype_t::ghost) {
       // Ghosts starts from local to end
-      return make_ids<index_space::entities>(util::iota_view<util::id>(
-        meta_field->local.ents, e_keys.span().size()));
+      return make_ids<index_space::entities>(
+        util::iota_view<util::id>(mf->local.ents, mf->local.ents + mf->ghosts));
     }
     else {
       // Iterate on all
@@ -1393,16 +1384,12 @@ public:
     traversal(
       &hmap.at(key_t::root()),
       [&](hcell_t * cur) {
-        if(cur->is_node()) {
+        if(cur->is_node())
           return Policy::intersect(
             e_i(ent_id), n_i(topo::id<ntree_base::nodes>(cur->node_idx())));
-        }
-        else {
-          if(Policy::intersect(e_i(ent_id),
-               e_i(topo::id<ntree_base::entities>(cur->ent_idx())))) {
-            ids.push_back(topo::id<ntree_base::entities>(cur->ent_idx()));
-          }
-        }
+        else if(Policy::intersect(e_i(ent_id),
+                  e_i(topo::id<ntree_base::entities>(cur->ent_idx()))))
+          ids.push_back(topo::id<ntree_base::entities>(cur->ent_idx()));
         return false;
       },
       hmap);
@@ -1412,21 +1399,17 @@ public:
   /// Return a range of all nodes of a \c ntree_base::ptype_t
   template<ptype_t PT = ptype_t::exclusive>
   auto nodes() const {
-    if constexpr(PT == ptype_t::exclusive) {
+    if constexpr(PT == ptype_t::exclusive)
       return make_ids<index_space::nodes>(
-        util::iota_view<util::id>(0, meta_field->local.nodes));
-    }
-    else if constexpr(PT == ptype_t::ghost) {
+        util::iota_view<util::id>(0, mf->local.nodes));
+    else if constexpr(PT == ptype_t::ghost)
       // Ghosts starts from local to end
-      return make_ids<index_space::entities>(
-        util::iota_view<util::id>(meta_field->local.nodes,
-          meta_field->local.nodes + meta_field->top_tree));
-    }
-    else {
-      // Iterate on all
       return make_ids<index_space::entities>(util::iota_view<util::id>(
-        0, meta_field->local.nodes + meta_field->top_tree));
-    }
+        mf->local.nodes, mf->local.nodes + mf->top_tree));
+    else
+      // Iterate on all
+      return make_ids<index_space::entities>(
+        util::iota_view<util::id>(0, mf->local.nodes + mf->top_tree));
   }
 
   /// Get nodes belonging to a node.
@@ -1553,18 +1536,18 @@ public:
 
 #if defined(FLECSI_ENABLE_GRAPHVIZ)
   /// Output a representation of the ntree using graphviz.
-  /// The output files are formatted as: colors()_tag.gv
+  /// The output files are formatted as: color()_tag.gv
   /// \param tag Tag for these files names
   void graphviz_draw(const std::string & tag) const {
     util::graphviz gv("G");
     static constexpr std::pair<const char *, const char *> completeness[] = {
-      {"incomplete", "doubleoctagon"},
-      { "complete",
-        "octagon" }};
+      {"complete", "octagon"},
+      { "incomplete",
+        "doubleoctagon" }};
     static constexpr std::pair<const char *, const char *> locality[] = {
-      {"local", "blue"},
-      { "non_local",
-        "red" }};
+      {"non_local", "red"},
+      { "local",
+        "blue" }};
 
     // Print a legend
     for(bool i : {true, false}) {
@@ -1661,6 +1644,9 @@ struct ntree_specialization : specialization<ntree, ntree_specialization> {
   /// Maximum number of neighbors per entities. This is used to compute
   /// neighbors list on GPU architectures.
   constexpr static unsigned int max_neighbors = 42;
+  /// Mandatory index spaces featured by the N-Tree interface.
+  /// This allows the access to the two index spaces entities and nodes.
+  using index_spaces = base::index_spaces;
 
   /// \name Intersection Functions
   /// Function computing interation between entity-entity, entity-node and